requirements.txt
aiosqlite==0.21.0
alembic==1.15.1
asyncpg==0.30.0
bcrypt==4.0.1
debugpy==1.8.1
fastapi[standard]==0.115.11
PyJWT==2.10.1
passlib==1.7.4
pytest-asyncio==0.26.0
pytest-lazy-fixtures==1.1.2
pydantic-settings==2.8.1
python-slugify==8.0.4
SQLAlchemy==2.0.38